# backend/docs/openapi.yaml
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
    FirebaseAuth:
      flows:
        implicit:
          authorizationUrl: https://accounts.google.com/o/oauth2/auth
          scopes:
            email: Email access
            openid: OpenID Connect
            profile: Profile access
      type: oauth2
info:
  contact:
    email: support@gaitanalysis.com
    name: Gait Analysis API Support
  description: "\n    ## Enhanced MediaPipe歩行分析API v2.0\n\n    スマートフォンで撮影した動画から歩行パターンを分析し、\n\
    \    理学療法士や医療従事者向けの詳細なレポートを生成する拡張APIです。\n\n    ### \U0001F680 新機能（v2.0 Enhanced）\n\
    \    - **30秒高速処理**: 撮影からレポート出力まで30秒以内\n    - **33ランドマーク活用**: MediaPipe Pose v0.10の全機能\n\
    \    - **代償動作検出**: 自動アラート機能付き\n    - **重心・バランス分析**: COM軌跡とバランス安定性評価\n    - **拡張出力**:\
    \ 正規化波形グラフ、スローモーション動画、カラーPDFレポート\n    - **iPad/iPhone最適化**: 60fps対応、横・後方撮影モード\n\
    \n    ### \U0001F4CA 詳細分析機能\n    - **スパイオテンポラル指標**: ピッチ、歩幅/ストライド長、立脚/遊脚時間、対称性指標\n\
    \    - **キネマティクス**: 股関節ROM、膝関節詳細角度、足関節背屈/底屈、骨盤・体幹回旋\n    - **代償動作アラート**: アウト/イントゥーイング、膝外反/内反、トレンデレンブルグ、ペルビックドロップ\n\
    \    - **バランス指標**: COM水平/垂直移動幅、重心投影線2Dマップ\n\n    ### \U0001F3AF 対象ユーザー\n    -\
    \ **理学療法士**: 詳細な歩行評価と治療計画立案\n    - **作業療法士**: 日常生活動作の評価\n    - **医師**: 診断支援と経過観察\n\
    \    - **研究者**: 歩行データの定量的解析\n    - **アスリート・トレーナー**: パフォーマンス向上\n\n    ### ⚙️ 技術仕様（Enhanced）\n\
    \    - **バックエンド**: FastAPI + Python 3.11+\n    - **AI処理**: MediaPipe Pose v0.10\
    \ + 最高精度モデル\n    - **並列処理**: マルチコア活用による高速化\n    - **データベース**: PostgreSQL + Redis\
    \ キャッシュ\n    - **認証**: Firebase Auth + JWT\n    - **監視**: リアルタイムパフォーマンス監視\n  \
    \  - **出力**: PDF/MP4/JSON/CSV 多形式対応\n    "
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  title: Enhanced MediaPipe Gait Analysis API
  version: 2.0.0-enhanced
openapi: 3.1.0
paths:
...
  /api/v1/performance/alerts:
    get:
      description: パフォーマンスアラートを取得
      operationId: get_performance_alerts_api_v1_performance_alerts_get
      parameters:
      - description: 取得する時間範囲（時間）
        in: query
        name: hours
        required: false
        schema:
          default: 24
          description: 取得する時間範囲（時間）
          title: Hours
          type: integer
      - description: アラートの重要度フィルタ
        in: query
        name: severity
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: アラートの重要度フィルタ
          title: Severity
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Performance Alerts
      tags:
      - performance
...
  /api/v1/performance/optimization/apply:
    post:
      description: 最適化を適用
      operationId: apply_optimization_api_v1_performance_optimization_apply_post
      parameters:
      - in: query
        name: optimization_type
        required: true
        schema:
          title: Optimization Type
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Apply Optimization
      tags:
      - performance
...
servers:
...